+++ packages/core/src/index.ts
@@ -6,10 +6,14 @@ import { useTableContext } from "./table/table_context";
 import { useColContext } from "./table/cols/col_context";
 import { VirtualHeader } from "./table/cols/virtual_header/types";
 import { useTableCssVars } from "./skin";
-import { defaultSkin, darkModeVars, lightModeVars } from "./default_skin/default_skin";
+import {
+  defaultSkin,
+  darkModeVars,
+  lightModeVars,
+} from "./default_skin/default_skin";
 import { useVirtualHeader } from "./table/cols/virtual_header/context";
 import { useRow } from "./table/rows/row_context";
-
+import { useMeasureHeader } from "./use_measure_header";
 export {
   ReactTanstackTableUi,
   useVirtualRowContext,
@@ -22,5 +26,6 @@ export {
   darkModeVars,
   lightModeVars,
   useVirtualHeader,
+  useMeasureHeader,
 };
 export type { Skin, VirtualHeader };